animeta/requirements.txt
Django==1.2.4
South==0.7.2
pytz
tweepy==1.7.1
-e hg+https://bitbucket.org/ditto/django-oauth-provider#egg=django-oauth
oauth2
facebook-python-sdk